+++ spec/lib/polish_geeks/dev_tools/validators/simplecov_spec.rb
@@ -1,12 +1,23 @@
 require 'spec_helper'
 
 RSpec.describe PolishGeeks::DevTools::Validators::Simplecov do
-  let(:output) { double(rspec: '370 / 669 LOC (95.00%) covered') }
+  let(:output) do
+    OpenStruct.new(
+      rspec: '370 / 669 LOC (95.00%) covered'
+    )
+  end
   subject { described_class.new(output) }
 
   describe '#valid?' do
+    subject do
+      described_class.new(
+        OpenStruct.new(
+          rspec: 'output'
+        )
+      )
+    end
+
     context 'false' do
-      subject { described_class.new(double(rspec: 'output')) }
       context 'when SimpleCov not defined' do
         before { allow(Object).to receive(:const_defined?) { false } }
         it do
@@ -17,7 +28,6 @@ RSpec.describe PolishGeeks::DevTools::Validators::Simplecov do
     end
 
     context 'true' do
-      subject { described_class.new(double(rspec: 'output')) }
       context 'when SimpleCov defined' do
         before { allow(Object).to receive(:const_defined?) { true } }
         it do